feat: Update EcommerceGoogleShopping struct to include timeout parameter
The EcommerceGoogleShopping struct has been updated to include a new timeout parameter. This parameter allows the user to specify the maximum timeout for the HTTP client when making requests to the Google Shopping API. The value is passed to the NewEcommerceGoogleShopping function and stored in the timeout field of the struct.